duration of each predicted trade
WM stands for WAITS-META, which refers to the time it took for every trade to finish (buy, then sell to target). used yFinance API to look up the differential between the start of the prediction (theoretical purchase time), then get the time when the price reached a 1% gain. months of june through july 5th '24